ISSUE #2489 - move multer/req code to routes
